+++ src/DomCrawler/Field/ChoiceFormField.php
@@ -25,15 +25,8 @@ final class ChoiceFormField extends BaseChoiceFormField
 {
     use FormFieldTrait;
 
-    /**
-     * @var string
-     */
-    private $type;
-
-    /**
-     * @var WebDriverSelectInterface
-     */
-    private $selector;
+    private string $type;
+    private WebDriverSelectInterface $selector;
 
     public function hasValue(): bool
     {
@@ -79,7 +72,7 @@ final class ChoiceFormField extends BaseChoiceFormField
         $this->setValue(false);
     }
 
-    public function getValue()
+    public function getValue(): array|string|null
     {
         $type = $this->element->getAttribute('type');
 
@@ -182,7 +175,7 @@ final class ChoiceFormField extends BaseChoiceFormField
     /**
      * Disables the internal validation of the field.
      */
-    public function disableValidation(): self
+    public function disableValidation(): static
     {
         throw $this->createNotSupportedException(__METHOD__);
     }