+++ src/DomCrawler/Link.php
@@ -24,7 +24,7 @@ final class Link extends BaseLink
 {
     use ExceptionThrower;
 
-    private $element;
+    private WebDriverElement $element;
 
     public function __construct(WebDriverElement $element, string $currentUri)
     {
@@ -43,7 +43,7 @@ final class Link extends BaseLink
         return $this->element;
     }
 
-    public function getNode()
+    public function getNode(): \DOMElement
     {
         throw $this->createNotSupportedException(__METHOD__);
     }